+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
 * Unreleased
-    * Add `EpoxyMockDigitalWriteFast`, a simple mock library for
+* 0.7 (2021-04-28)
+    * Add `libraries/EpoxyMockDigitalWriteFast`, a simple mock library for
       `digitalWriteFast` (https://github.com/NicksonYap/digitalWriteFast) to
       allow code using these functions (e.g. `digitalWriteFast()`) to compile
       under EpoxyDuino.
-    * Rename `EpoxyPromAvr` to `EpoxyEepromAvr`; `EpoxyPromEsp` to
-      `EpoxyEepromEsp` for consistency and better self-description.
+    * Rename `EpoxyPromAvr` to `EpoxyEepromAvr`, and `EpoxyPromEsp` to
+      `EpoxyEepromEsp`, for consistency and better self-description.
+    * Add `EPOXY_CORE` and `EPOXY_CORE_PATH` variables to allow alternate
+      Arduino Core files to be specified. Move current Core files from top-level
+      into `./cores/epoxy/` subdirectory.
+    * Add `memcpy_P()` and `vsnprintf_P()` (Thanks pmp-p@, #28).
 * 0.6.2 (2021-03-15)
     * Add more stubs for Arduino API functions: `pulseIn()`, `pulseInLong()`,
       `shiftOut()`, `shiftIn()`, I2C and SPI pins (SS, MOSI, MISO, SCK, SDA,